+++ app/javascript/packs/application.js
@@ -33,7 +33,6 @@ document.addEventListener('turbolinks:load', () => {
   // Call your functions here, e.g:
   // initSelect2();
   initLocation();
-  initTextParser();
 });
 
 import "controllers"
