+++ app/views/chomp_sessions/_form.html.erb
@@ -1,4 +1,7 @@
-<%= simple_form_for(chomp_session) do |f| %>
+<%= simple_form_for(chomp_session,
+                    data: {
+                      controller: "chomp"
+                    }) do |f| %>
   <%= f.input :name, label:"Meeting Name", required: false,
     placeholder: "I can't decide on a name on #{Time.now}",
     input_html: { value: f.object.name || [
@@ -15,9 +18,12 @@
     "Turtles are cute",
     "Cats > Dogs",
     "Dogs > Cats"
-    ].sample, id: "chomp-session-name" } %>
+    ].sample,
+    id: "chomp-session-name",
+    data: { chomp_target: "name", action: "keyup->chomp#autofillTime" }
+    } %>
   <%= f.input :date, html5: true, min: "#{Date.today.iso8601}"%>
-  <%= f.input :time, html5: true, min: "#{Time.now.iso8601}", input_html: { id: "chomp-session-time" } %>
+  <%= f.input :time, html5: true, min: "#{Time.now.iso8601}", input_html: { id: "chomp-session-time", data: {chomp_target: "timeOutput"} } %>
   <%= f.input :session_expiry, label:"Link Expiry Time (Hours)", hint: "Set the time that you want your poll to be active" %>
   <%# <%= f.input :invitees, label:"Number of Friends (optional)", placeholder:"Optional" %>
